Funcionalidades todas check Código HTML validado Código css validado

+++ task.js
@@ -5,20 +5,30 @@ const tasks = JSON.parse(localStorage.getItem("tasks"));
 const task = JSON.parse(localStorage.getItem("task_object"));
 const index = localStorage.getItem("task_index");
 
-document.querySelector("#scrum_color").style.backgroundColor = localStorage.getItem("background_color");
+document.querySelector("#body_color").style.backgroundColor = localStorage.getItem("background_color");
 
 console.log(tasks);
 
-document.querySelector("#user_task").textContent = localStorage.getItem("username");
+document.querySelector("#user").textContent = localStorage.getItem("username");
+
+// executa a função imediatamente
+writeDate();
+
+// executa a função em intervalos de 1 segundo para atualizar a data
+setInterval(writeDate, 1000);
 
 if (task.title != "") {
    title_txt.value = task.title;
    description_txt.value = task.description;
    color_input.value = task.color;
+   document.querySelector("#task_save").style.width = "46%";
    document.querySelector("#task_delete").style.display = "inline-block";
+   document.querySelector("#task_creationTitle").textContent = "Task Edit";
 } else {
    color_input.value = "#ffffff";
    document.querySelector("#task_delete").style.display = "none";
+   document.querySelector("#task_save").style.width = "95%";
+   document.querySelector("#task_creationTitle").textContent = "Task Creation";
 }
 
 document.querySelector("#task_save").addEventListener("click", function () {
@@ -36,11 +46,13 @@ document.querySelector("#task_save").addEventListener("click", function () {
          localStorage.setItem("tasks", JSON.stringify(tasks));
          window.location.href = "scrum.html";
       } else {
-         tasks[index].title = title_txt.value;
-         tasks[index].description = description_txt.value;
-         tasks[index].color = color_input.value;
-         localStorage.setItem("tasks", JSON.stringify(tasks));
-         window.location.href = "scrum.html";
+         if (confirmEdit()) {
+            tasks[index].title = title_txt.value;
+            tasks[index].description = description_txt.value;
+            tasks[index].color = color_input.value;
+            localStorage.setItem("tasks", JSON.stringify(tasks));
+            window.location.href = "scrum.html";
+         }
       }
    } else {
       alert("Need to put a task title.");
@@ -48,11 +60,48 @@ document.querySelector("#task_save").addEventListener("click", function () {
 });
 
 document.querySelector("#task_delete").addEventListener("click", function () {
-   tasks.splice(index, 1);
-   localStorage.setItem("tasks", JSON.stringify(tasks));
-   window.location.href = "scrum.html";
+   if (confirmDelete()) {
+      tasks.splice(index, 1);
+      localStorage.setItem("tasks", JSON.stringify(tasks));
+      window.location.href = "scrum.html";
+   }
 });
 
 document.querySelector("#cancel").addEventListener("click", function () {
-   window.location.href = "scrum.html";
+   if (confirmExit()) {
+      window.location.href = "scrum.html";
+   }
 });
+
+document.querySelector("#logout").addEventListener("click", function () {
+   localStorage.setItem("username", "");
+   window.location.href = "login.html";
+});
+
+//Função para confirmar delete
+function confirmDelete() {
+   return confirm("Are you sure you want to delete this?");
+}
+
+//Função para confirmar edit
+function confirmEdit() {
+   return confirm("Are you sure you want to edit this?");
+}
+
+//Função para confirmar sair da janela
+function confirmExit() {
+   return confirm("Are you sure you want to exit without saving first?");
+}
+
+// Função data e relógio
+
+function writeDate() {
+   const d = new Date();
+
+   // define o formato a mostrar
+   let dateTimeString = d.toLocaleString("en-GB");
+   dateTimeString = dateTimeString.replace(",", "&nbsp; &nbsp; &nbsp;");
+
+   // insere no HTML
+   document.getElementById("date").innerHTML = dateTimeString;
+}
